+++ playwright.config.ts
@@ -5,17 +5,41 @@ export default defineConfig({
   fullyParallel: true,
   reporter: [["list"], ["html", { open: "never" }]],
   use: {
-    baseURL: process.env.BASE_URL || "http://localhost:6006",
     headless: true,
     trace: "retain-on-failure",
     screenshot: "on",
     video: "on-first-retry",
   },
-  webServer: {
-    command: "npm run storybook -- --ci",
-    port: 6006,
-    reuseExistingServer: !process.env.CI,
-    stdout: "ignore",
-    stderr: "pipe",
-  },
+  projects: [
+    {
+      name: "component",
+      use: {
+        baseURL: "http://localhost:6006",
+      },
+      testMatch: /.*\.component\.spec\.ts/,
+    },
+    {
+      name: "integration",
+      use: {
+        baseURL: "http://localhost:3000",
+      },
+      testMatch: /(routing|example)\.spec\.ts/,
+    },
+  ],
+  webServer: [
+    {
+      command: "npm run start",
+      port: 3000,
+      reuseExistingServer: !process.env.CI,
+      stdout: "ignore",
+      stderr: "pipe",
+    },
+    {
+      command: "npm run storybook -- --ci",
+      port: 6006,
+      reuseExistingServer: !process.env.CI,
+      stdout: "ignore",
+      stderr: "pipe",
+    },
+  ],
 });